Test the OMEdit diagrams
Renamed RunOMEditTestsuite to RunOMEditTestsuite.sh
Merged the browsing tests
