+++ OMEdit/Testsuite/Makefile.omdev.mingw
@@ -22,7 +22,7 @@ mkbuilddirs:
 
 install: build
 	cp -p ../bin/tests/*$(EXE) $(builddir_bin)
-	cp -p RunOMEditTestsuite $(builddir_bin)
+	cp -p RunOMEditTestsuite.sh $(builddir_bin)
 
 build: Makefile
 	$(MAKE) -f Makefile
