+++ OMEdit/Testsuite/Makefile.unix.in
@@ -4,7 +4,7 @@ QMAKE=@QMAKE@
 
 install: build
 	cp -p ../bin/tests/* @OMBUILDDIR@/bin
-	cp -p RunOMEditTestsuite @OMBUILDDIR@/bin
+	cp -p RunOMEditTestsuite.sh @OMBUILDDIR@/bin
 
 build: Makefile OMEditGUI.unix.config.pri
 	$(MAKE) -f Makefile
